move rules to separate file
