+++ src/insert/spacefinder/article.ts
@@ -7,62 +7,19 @@ import {
 	wrapSlotInContainer,
 } from 'core/create-ad-slot';
 import { spaceFiller } from 'insert/spacefinder/space-filler';
-import type {
-	RuleSpacing,
-	SpacefinderRules,
-	SpacefinderWriter,
-} from 'insert/spacefinder/spacefinder';
+import type { SpacefinderWriter } from 'insert/spacefinder/spacefinder';
 import { commercialFeatures } from 'lib/commercial-features';
-import {
-	getCurrentBreakpoint,
-	getCurrentTweakpoint,
-} from 'lib/detect/detect-breakpoint';
+import { getCurrentBreakpoint } from 'lib/detect/detect-breakpoint';
 import fastdom from '../../utils/fastdom-promise';
 import { computeStickyHeights, insertHeightStyles } from '../sticky-inlines';
 import { initCarrot } from './carrot-traffic-driver';
-import { isInHighValueSection } from './utils';
+import { rules } from './rules';
 
 type SlotName = Parameters<typeof createAdSlot>[0];
-
-/**
- * As estimation of the height of the most viewed island.
- * This appears from desktop breakpoints on the right-hand side.
- * Knowing the height of the element is useful when
- * calculating where to place ads in the right column.
- */
-const MOST_VIEWED_HEIGHT = 600;
 
 const articleBodySelector = '.article-body-commercial-selector';
 
 const isPaidContent = window.guardian.config.page.isPaidContent;
-
-const hasImages = !!window.guardian.config.page.lightboxImages?.images.length;
-
-const hasShowcaseMainElement =
-	window.guardian.config.page.hasShowcaseMainElement;
-
-const minDistanceBetweenRightRailAds = 500;
-const minDistanceBetweenInlineAds = isInHighValueSection ? 500 : 750;
-
-/**
- * Rules to avoid inserting ads in the right rail too close to each other
- **/
-const rightRailAdSlotContainerRules: Record<string, RuleSpacing> = {
-	[` .${adSlotContainerClass}`]: {
-		minAboveSlot: minDistanceBetweenRightRailAds,
-		minBelowSlot: minDistanceBetweenRightRailAds,
-	},
-};
-
-/**
- * Rules to avoid inserting inline ads too close to each other
- **/
-const inlineAdSlotContainerRules: Record<string, RuleSpacing> = {
-	[` .${adSlotContainerClass}`]: {
-		minAboveSlot: minDistanceBetweenInlineAds,
-		minBelowSlot: minDistanceBetweenInlineAds,
-	},
-};
 
 /**
  * Get the classname for an ad slot container
@@ -143,47 +100,6 @@ const decideAdditionalSizes = async (
 };
 
 const addDesktopInline1 = (fillSlot: FillAdSlot): Promise<boolean> => {
-	const tweakpoint = getCurrentTweakpoint();
-	const hasLeftCol = ['leftCol', 'wide'].includes(tweakpoint);
-
-	let ignoreList = `:scope > :not(p):not(h2):not(ul):not(.${adSlotContainerClass}):not(#sign-in-gate):not(.sfdebug)`;
-	if (hasLeftCol) {
-		ignoreList +=
-			':not([data-spacefinder-role="richLink"]):not([data-spacefinder-role="thumbnail"])';
-	}
-
-	const isImmersive = window.guardian.config.page.isImmersive;
-
-	const rules: SpacefinderRules = {
-		bodySelector: articleBodySelector,
-		candidateSelector: ':scope > p, [data-spacefinder-role="nested"] > p',
-		minAbove: isImmersive ? 700 : 300,
-		minBelow: 300,
-		opponentSelectorRules: {
-			// don't place ads right after a heading
-			':scope > h2, [data-spacefinder-role="nested"] > h2': {
-				minAboveSlot: isInHighValueSection ? 150 : 5,
-				minBelowSlot: isInHighValueSection ? 0 : 190,
-			},
-			[`.${adSlotContainerClass}`]: {
-				minAboveSlot: 500,
-				minBelowSlot: 500,
-			},
-			[ignoreList]: {
-				minAboveSlot: 35,
-				minBelowSlot: 400,
-			},
-			'[data-spacefinder-role="immersive"]': {
-				minAboveSlot: 0,
-				minBelowSlot: 600,
-			},
-			'figure.element--supporting': {
-				minAboveSlot: 100,
-				minBelowSlot: 0,
-			},
-		},
-	};
-
 	// these are added here and not in size mappings because the inline[i] name
 	// is also used on fronts, where we don't want outstream or tall ads
 	const additionalSizes = {
@@ -201,7 +117,7 @@ const addDesktopInline1 = (fillSlot: FillAdSlot): Promise<boolean> => {
 		await Promise.all(slots);
 	};
 
-	return spaceFiller.fillSpace(rules, insertAd, {
+	return spaceFiller.fillSpace(rules.desktopInline1, insertAd, {
 		waitForImages: true,
 		waitForInteractives: true,
 		pass: 'inline1',
@@ -212,51 +128,6 @@ const addDesktopInline1 = (fillSlot: FillAdSlot): Promise<boolean> => {
  * Inserts all inline ads on desktop except for inline1.
  */
 const addDesktopRightRailAds = (fillSlot: FillAdSlot): Promise<boolean> => {
-	let minAbove = 1000;
-
-	/**
-	 * In special cases, inline2 can overlap the "Most viewed" island, so
-	 * we need to make an adjustment to move the inline2 further down the page
-	 */
-	if (isPaidContent) {
-		minAbove += MOST_VIEWED_HEIGHT;
-	}
-
-	// Some old articles don't have a main image, which means the first paragraph is much higher
-	if (!hasImages) {
-		minAbove += 600;
-	} else if (hasShowcaseMainElement) {
-		minAbove += 100;
-	}
-
-	const largestSizeForSlot = adSizes.halfPage.height;
-	const rules: SpacefinderRules = {
-		bodySelector: articleBodySelector,
-		candidateSelector: ':scope > p, [data-spacefinder-role="nested"] > p',
-		minAbove,
-		minBelow: 300,
-		opponentSelectorRules: {
-			...rightRailAdSlotContainerRules,
-			'[data-spacefinder-role="immersive"]': {
-				minAboveSlot: 0,
-				minBelowSlot: 600,
-			},
-		},
-		/**
-		 * Filter out any candidates that are too close to the last winner
-		 * see https://github.com/guardian/commercial/tree/main/docs/spacefinder#avoiding-other-winning-candidates
-		 * for more information
-		 **/
-		filter: (candidate, lastWinner) => {
-			if (!lastWinner) {
-				return true;
-			}
-			const distanceBetweenAds =
-				candidate.top - lastWinner.top - largestSizeForSlot;
-			return distanceBetweenAds >= minDistanceBetweenRightRailAds;
-		},
-	};
-
 	const insertAds: SpacefinderWriter = async (paras) => {
 		const stickyContainerHeights = await computeStickyHeights(
 			paras,
@@ -306,46 +177,14 @@ const addDesktopRightRailAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 		await Promise.all(slots);
 	};
 
-	return spaceFiller.fillSpace(rules, insertAds, {
+	return spaceFiller.fillSpace(rules.desktopRightRail, insertAds, {
 		waitForImages: true,
 		waitForInteractives: true,
 		pass: 'subsequent-inlines',
 	});
 };
 
-const mobileMinDistanceFromArticleTop = 200;
-
-const mobileIgnoreList = `:not(p):not(h2):not(hr):not(.${adSlotContainerClass}):not(#sign-in-gate):not([data-spacefinder-type$="NumberedTitleBlockElement"])`;
-
-const mobileOpponentSelectorRules = {
-	// don't place ads right after a heading
-	':scope > h2, [data-spacefinder-role="nested"] > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"]':
-		{
-			minAboveSlot: 100,
-			minBelowSlot: 0,
-		},
-	...inlineAdSlotContainerRules,
-	// this is a catch-all for elements that are not covered by the above rules, these will generally be things like videos, embeds and atoms. minBelowSlot is higher to push ads a bit further down after these elements
-	[`:scope > ${mobileIgnoreList}, [data-spacefinder-role="nested"] > ${mobileIgnoreList}`]:
-		{
-			minAboveSlot: 35,
-			minBelowSlot: 200,
-			// Usually we don't want an ad right before videos, embeds and atoms etc. so that we don't break up related content too much. But if we have a heading above, anything above the heading won't be related to the current content, so we can place an ad there.
-			bypassMinBelow:
-				'h2,[data-spacefinder-type$="NumberedTitleBlockElement"]',
-		},
-};
-
 const addMobileTopAboveNav = (fillSlot: FillAdSlot): Promise<boolean> => {
-	const rules: SpacefinderRules = {
-		bodySelector: articleBodySelector,
-		candidateSelector:
-			':scope > p, :scope > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"], [data-spacefinder-role="nested"] > p',
-		minAbove: mobileMinDistanceFromArticleTop,
-		minBelow: 200,
-		opponentSelectorRules: mobileOpponentSelectorRules,
-	};
-
 	const insertAds: SpacefinderWriter = async (paras) => {
 		const slots = paras.slice(0, 1).map(async (para) => {
 			const name = 'top-above-nav';
@@ -356,7 +195,7 @@ const addMobileTopAboveNav = (fillSlot: FillAdSlot): Promise<boolean> => {
 		await Promise.all(slots);
 	};
 
-	return spaceFiller.fillSpace(rules, insertAds, {
+	return spaceFiller.fillSpace(rules.mobileTopAboveNav, insertAds, {
 		waitForImages: true,
 		waitForInteractives: true,
 		pass: 'mobile-top-above-nav',
@@ -366,27 +205,6 @@ const addMobileTopAboveNav = (fillSlot: FillAdSlot): Promise<boolean> => {
 const addMobileSubsequentInlineAds = (
 	fillSlot: FillAdSlot,
 ): Promise<boolean> => {
-	const rules: SpacefinderRules = {
-		bodySelector: articleBodySelector,
-		candidateSelector:
-			':scope > p, :scope > h2, :scope > [data-spacefinder-type$="NumberedTitleBlockElement"], [data-spacefinder-role="nested"] > p',
-		minAbove: mobileMinDistanceFromArticleTop,
-		minBelow: 200,
-		opponentSelectorRules: mobileOpponentSelectorRules,
-		/**
-		 * Filter out any candidates that are too close to the last winner
-		 * see https://github.com/guardian/commercial/tree/main/docs/spacefinder#avoiding-other-winning-candidates
-		 * for more information
-		 **/
-		filter: (candidate, lastWinner) => {
-			if (!lastWinner) {
-				return true;
-			}
-			const distanceBetweenAds = candidate.top - lastWinner.top;
-			return distanceBetweenAds >= minDistanceBetweenInlineAds;
-		},
-	};
-
 	const insertAds: SpacefinderWriter = async (paras) => {
 		const slots = paras.map(async (para, i) => {
 			const name = `inline${i + 1}`;
@@ -406,7 +224,7 @@ const addMobileSubsequentInlineAds = (
 		await Promise.all(slots);
 	};
 
-	return spaceFiller.fillSpace(rules, insertAds, {
+	return spaceFiller.fillSpace(rules.mobileSubsequentInlineAds, insertAds, {
 		waitForImages: true,
 		waitForInteractives: true,
 		pass: 'mobile-subsequent-inlines',
@@ -443,40 +261,6 @@ const addSubsequentInlineAds = (fillSlot: FillAdSlot): Promise<boolean> => {
 const attemptToAddInlineMerchAd = (
 	fillAdSlot: FillAdSlot,
 ): Promise<boolean> => {
-	const breakpoint = getCurrentBreakpoint();
-	const isMobileOrTablet = breakpoint === 'mobile' || breakpoint === 'tablet';
-
-	const rules: SpacefinderRules = {
-		bodySelector: articleBodySelector,
-		candidateSelector: ':scope > p',
-		minAbove: 300,
-		minBelow: 300,
-		opponentSelectorRules: {
-			':scope > .merch': {
-				minAboveSlot: 0,
-				minBelowSlot: 0,
-			},
-			':scope > header': {
-				minAboveSlot: isMobileOrTablet ? 300 : 700,
-				minBelowSlot: 0,
-			},
-			':scope > h2': {
-				minAboveSlot: 100,
-				minBelowSlot: 250,
-			},
-			':scope > #sign-in-gate': {
-				minAboveSlot: 0,
-				minBelowSlot: 400,
-			},
-			...inlineAdSlotContainerRules,
-			[`:scope > :not(p):not(h2):not(.${adSlotContainerClass}):not(#sign-in-gate):not(.sfdebug)`]:
-				{
-					minAboveSlot: 200,
-					minBelowSlot: 400,
-				},
-		},
-	};
-
 	const insertAds: SpacefinderWriter = async (paras) => {
 		if (typeof paras[0] === 'undefined') {
 			throw new Error(
@@ -490,7 +274,7 @@ const attemptToAddInlineMerchAd = (
 		await fillAdSlot('im', slot);
 	};
 
-	return spaceFiller.fillSpace(rules, insertAds, {
+	return spaceFiller.fillSpace(rules.inlineMerchandising, insertAds, {
 		waitForImages: true,
 		waitForInteractives: true,
 		pass: 'im',
